Have LegacyMessage and the new MessageBundles side by side
