+++ inlang/source-code/versioned-interfaces/message/src/interface.ts
@@ -1,38 +1,19 @@
 import { LanguageTag } from "@inlang/language-tag"
 import { Type, type Static } from "@sinclair/typebox"
 
-export type Literal = Static<typeof Literal>
-export const Literal = Type.Object({
-	type: Type.Literal("literal"),
-	value: Type.String(),
-})
-
 /**
  * A (text) element that is translatable and rendered to the UI.
  */
 export type Text = Static<typeof Text>
 export const Text = Type.Object({
-	type: Type.Literal("text"),
+	type: Type.Literal("Text"),
 	value: Type.String(),
 })
 
 export type VariableReference = Static<typeof VariableReference>
 export const VariableReference = Type.Object({
-	type: Type.Literal("variable"),
+	type: Type.Literal("VariableReference"),
 	name: Type.String(),
-})
-
-export type Option = Static<typeof Option>
-export const Option = Type.Object({
-	name: Type.String(),
-	value: Type.Union([Literal, VariableReference]),
-})
-
-export type FunctionAnnotation = Static<typeof FunctionAnnotation>
-export const FunctionAnnotation = Type.Object({
-	type: Type.Literal("function"),
-	name: Type.String(),
-	options: Type.Array(Option),
 })
 
 /**
@@ -42,11 +23,7 @@ export const FunctionAnnotation = Type.Object({
  * text value during runtime.
  */
 export type Expression = Static<typeof Expression>
-export const Expression = Type.Object({
-	type: Type.Literal("expression"),
-	arg: Type.Union([Literal, VariableReference]),
-	annotation: Type.Optional(FunctionAnnotation),
-})
+export const Expression = Type.Union([VariableReference])
 
 // export type FunctionReference = {
 // 	type: "function"
@@ -67,6 +44,7 @@ export const Pattern = Type.Array(Type.Union([Text, Expression]))
  */
 export type Variant = Static<typeof Variant>
 export const Variant = Type.Object({
+	languageTag: LanguageTag,
 	/**
 	 * The number of keys in each variant match MUST equal the number of expressions in the selectors.
 	 *
@@ -77,44 +55,13 @@ export const Variant = Type.Object({
 	pattern: Pattern,
 })
 
-export type InputDeclaration = Static<typeof InputDeclaration>
-export const InputDeclaration = Type.Object({
-	type: Type.Literal("input"),
-	name: Type.String(),
-
-	//TODO make this generic so that only Variable-Ref Expressions are allowed
-	value: Expression,
-})
-
-export type LocalDeclaration = Static<typeof InputDeclaration>
-export const LocalDeclaration = Type.Object({
-	type: Type.Literal("local"),
-	name: Type.String(),
-	value: Expression,
-})
-
-export type Declaration = Static<typeof Declaration>
-export const Declaration = Type.Union([LocalDeclaration, InputDeclaration])
-
-export type Translation = Static<typeof Translation>
-export const Translation = Type.Object({
-	languageTag: LanguageTag,
-	declarations: Type.Array(Declaration),
+export type Message = Static<typeof Message>
+export const Message = Type.Object({
+	id: Type.String(),
+	alias: Type.Record(Type.String(), Type.String()),
 	/**
 	 * The order in which the selectors are placed determines the precedence of patterns.
 	 */
 	selectors: Type.Array(Expression),
 	variants: Type.Array(Variant),
-})
-
-export type Message = Static<typeof Message>
-export const Message = Type.Object({
-	id: Type.String(),
-	alias: Type.Record(Type.String(), Type.String()),
-
-	/**
-	 * The canonical order for named inputs
-	 */
-	inputs: Type.Array(Type.String()),
-	translations: Type.Array(Translation),
 })
